update blogpost cerita
